+++ src/services/firestore/mutation.ts
@@ -1,0 +1,5 @@
+export default `mutation($input: [AddgcpFirestoreDatabaseInput!]!) {
+  addgcpFirestoreDatabase(input: $input, upsert: true) {
+    numUids
+  }
+}`;