+++ src/types/generated.ts
@@ -30,10 +30,10 @@ export type GcpAlertPolicy = GcpBaseResource & {
   displayName?: Maybe<Scalars['String']>;
   documentation?: Maybe<GcpAlertPolicyDocumentation>;
   enabled?: Maybe<GcpBoolValue>;
+  labels?: Maybe<Array<Maybe<GcpRawLabel>>>;
   mutationRecord?: Maybe<GcpAlertPolicyMutationRecord>;
   notificationChannels?: Maybe<Array<Maybe<Scalars['String']>>>;
   project?: Maybe<Array<Maybe<GcpProject>>>;
-  userLabels?: Maybe<Array<Maybe<GcpRawLabel>>>;
   validity?: Maybe<GcpRpcStatus>;
 };
 
@@ -1757,6 +1757,17 @@ export type GcpFileContentBuffer = {
   id: Scalars['String'];
 };
 
+export type GcpFirestoreDatabase = GcpBaseResource & {
+  appEngineIntegrationMode?: Maybe<Scalars['String']>;
+  concurrencyMode?: Maybe<Scalars['String']>;
+  createTime?: Maybe<Scalars['String']>;
+  keyPrefix?: Maybe<Scalars['String']>;
+  locationId?: Maybe<Scalars['String']>;
+  project?: Maybe<Array<Maybe<GcpProject>>>;
+  type?: Maybe<Scalars['String']>;
+  updateTime?: Maybe<Scalars['String']>;
+};
+
 export type GcpFirewall = {
   allowed?: Maybe<Array<Maybe<GcpFirewallAccess>>>;
   creationTimestamp?: Maybe<Scalars['String']>;
@@ -1923,6 +1934,20 @@ export type GcpKmsKeyRing = GcpBaseResource & {
   importJobs?: Maybe<Array<Maybe<GcpKmsImportJob>>>;
   kmsCryptoKeys?: Maybe<Array<Maybe<GcpKmsCryptoKey>>>;
   project?: Maybe<Array<Maybe<GcpProject>>>;
+};
+
+export type GcpLabel = GcpBaseResource & {
+  alertPolicy?: Maybe<Array<Maybe<GcpAlertPolicy>>>;
+  cloudFunction?: Maybe<Array<Maybe<GcpCloudFunction>>>;
+  dnsManagedZone?: Maybe<Array<Maybe<GcpDnsManagedZone>>>;
+  id: Scalars['String'];
+  key: Scalars['String'];
+  kmsCryptoKey?: Maybe<Array<Maybe<GcpKmsCryptoKey>>>;
+  project?: Maybe<Array<Maybe<GcpProject>>>;
+  secrets?: Maybe<Array<Maybe<GcpSecret>>>;
+  storageBucket?: Maybe<Array<Maybe<GcpStorageBucket>>>;
+  value: Scalars['String'];
+  vmInstance?: Maybe<Array<Maybe<GcpVmInstance>>>;
 };
 
 export type GcpLabelDescriptor = {
@@ -2136,6 +2161,7 @@ export type GcpProject = {
   dnsManagedZone?: Maybe<Array<Maybe<GcpDnsManagedZone>>>;
   dnsPolicy?: Maybe<Array<Maybe<GcpDnsPolicy>>>;
   etag?: Maybe<Scalars['String']>;
+  firestoreDatabases?: Maybe<Array<Maybe<GcpFirestoreDatabase>>>;
   firewall?: Maybe<Array<Maybe<GcpFirewall>>>;
   folder?: Maybe<Array<Maybe<GcpFolder>>>;
   iamPolicy?: Maybe<Array<Maybe<GcpIamPolicy>>>;
@@ -2152,7 +2178,7 @@ export type GcpProject = {
   organization?: Maybe<Array<Maybe<GcpOrganization>>>;
   parent?: Maybe<Scalars['String']>;
   projectId?: Maybe<Scalars['String']>;
-  secretManager?: Maybe<Array<Maybe<GcpSecret>>>;
+  secrets?: Maybe<Array<Maybe<GcpSecret>>>;
   serviceAccounts?: Maybe<Array<Maybe<GcpServiceAccount>>>;
   sqlInstances?: Maybe<Array<Maybe<GcpSqlInstance>>>;
   sslPolicies?: Maybe<Array<Maybe<GcpSslPolicy>>>;
@@ -2587,11 +2613,14 @@ export type GcpSubnetSecondaryRange = {
   rangeName?: Maybe<Scalars['String']>;
 };
 
-export type GcpTag = {
+export type GcpTag = GcpBaseResource & {
+  folder?: Maybe<Array<Maybe<GcpFolder>>>;
   id: Scalars['String'];
   key: Scalars['String'];
-  kms?: Maybe<Array<Maybe<GcpKmsKeyRing>>>;
+  organization?: Maybe<Array<Maybe<GcpOrganization>>>;
+  project?: Maybe<Array<Maybe<GcpProject>>>;
   value: Scalars['String'];
+  vmInstance?: Maybe<Array<Maybe<GcpVmInstance>>>;
 };
 
 export type GcpTargetHttpsProxy = GcpBaseResource & {